requirements.txt
amqp==2.2.2
billiard==3.5.0.3
cachetools==2.0.1
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
google-api-core==1.1.0
google-auth==1.4.1
google-cloud-core==0.28.1
google-cloud-storage==1.8.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
idna==2.6
kombu==4.1.0
numpy==1.14.2
pandas==0.22.0
protobuf==3.5.2.post1
psycopg2==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-dateutil==2.7.2
pytz==2018.3
redis==2.10.6
requests==2.18.4
rsa==3.4.2
six==1.11.0
SQLAlchemy==1.2.6
urllib3==1.22
vine==1.1.4
